test: extract MockEventBus
