+++ internal/project/application/usecases/shared_test.go
@@ -1,36 +1,1 @@
 package usecases_test
-
-import (
-	"os"
-
-	"github.com/stretchr/testify/mock"
-
-	"gomander/internal/eventbus"
-)
-
-type MockFsFacade struct {
-	mock.Mock
-}
-
-func (m *MockFsFacade) WriteFile(path string, data []byte, perm os.FileMode) error {
-	args := m.Called(path, data, perm)
-	return args.Error(0)
-}
-
-func (m *MockFsFacade) ReadFile(path string) ([]byte, error) {
-	args := m.Called(path)
-	return args.Get(0).([]byte), args.Error(1)
-}
-
-type MockEventBus struct {
-	mock.Mock
-}
-
-func (m *MockEventBus) RegisterHandler(handler eventbus.EventHandler) {
-	m.Called(handler)
-}
-
-func (m *MockEventBus) PublishSync(e eventbus.Event) []error {
-	args := m.Called(e)
-	return args.Get(0).([]error)
-}
